+++ core/common/src/main/java/caios/android/kanade/core/common/network/util/CoroutineUtils.kt
@@ -1,0 +1,13 @@
+package caios.android.kanade.core.common.network.util
+
+import timber.log.Timber
+import kotlin.coroutines.cancellation.CancellationException
+
+suspend fun <T> suspendRunCatching(block: suspend () -> T): Result<T> = try {
+    Result.success(block())
+} catch (cancellationException: CancellationException) {
+    throw cancellationException
+} catch (exception: Exception) {
+    Timber.i(exception, "Failed to evaluate a suspendRunCatchingBlock. Returning failure Result")
+    Result.failure(exception)
+}
