+++ core/model/src/main/java/caios/android/kanade/core/model/entity/YTMusicOAuthCode.kt
@@ -1,0 +1,18 @@
+package caios.android.kanade.core.model.entity
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class YTMusicOAuthCode(
+    @SerialName("device_code")
+    val deviceCode: String,
+    @SerialName("expires_in")
+    val expiresIn: Int,
+    @SerialName("interval")
+    val interval: Int,
+    @SerialName("user_code")
+    val userCode: String,
+    @SerialName("verification_url")
+    val verificationUrl: String
+)
